refactor(constants): add publications
