+++ src/constants/index.js
@@ -1,0 +1,3 @@
+import publications from './publications';
+
+export { publications };
